refactor: split db/src/hls.rs into hls/ subdirectory (#534)
Mechanical split of db/src/hls.rs (816 lines, over the 800-line cap)
into hls/{mod,fs,query,manifest,transcode}.rs per the suggested
approach. No behavior changes; the public surface is preserved via
pub use re-exports from mod.rs so call sites in server/ and
db/src/worker/handlers.rs are unchanged.

mod.rs holds the cross-zone items (HlsError, HlsPart, ResolvedAudiobook,
AUDIO64) and re-exports the public API. fs.rs hosts the cache directory
layout and the .progress/.failed sentinel helpers (plus the private
is_progress_stale_at, surfaced to transcode.rs via pub(super)). query.rs
holds the four sqlx audiobook lookups. manifest.rs is the pure
build_manifest + ffmpeg progress parse/fraction helpers. transcode.rs is
the ffmpeg engine: transcode_book, concat input prep, child spawn,
heartbeat reader, watchdog timeout, finalize, and FIFO-by-mtime
evict_if_over_cap.

The companion tests file (db/src/hls/tests.rs) keeps its `use super::*`
and now imports `std::time::{Duration, SystemTime}` explicitly since
those types no longer live in the parent module's import scope.

Closes #534
